Rename MediaNotificationFeature to MediaFeature.
