+++ components/feature/media/src/main/java/mozilla/components/feature/media/MediaFeature.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.media.notification
+package mozilla.components.feature.media
 
 import android.content.Context
 import mozilla.components.feature.media.service.MediaService
@@ -10,11 +10,16 @@ import mozilla.components.feature.media.state.MediaState
 import mozilla.components.feature.media.state.MediaStateMachine
 
 /**
- * Feature for displaying an ongoing notification (keeping the app process alive) while web content is playing media.
+ * Feature implementation for media playback in web content. This feature takes care of:
+ * - Background playback without the app getting killed.
+ * - Showing a media notification with play/pause controls.
+ * - Audio Focus handling (pausing/resuming in agreement with other media apps).
+ * - Support for hardware controls to toggle play/pause (e.g. buttons on a headset)
  *
- * This feature should get initialized globally once on app start.
+ * This feature should get initialized globally once on app start and requires a started
+ * [MediaStateMachine].
  */
-class MediaNotificationFeature(
+class MediaFeature(
     private val context: Context
 ) {
     private var serviceRunning = false
@@ -39,7 +44,7 @@ class MediaNotificationFeature(
 }
 
 internal class MediaObserver(
-    private val feature: MediaNotificationFeature
+    private val feature: MediaFeature
 ) : MediaStateMachine.Observer {
     override fun onStateChanged(state: MediaState) {
         if (state is MediaState.Playing || state is MediaState.Paused) {
