+++ components/feature/media/src/test/java/mozilla/components/feature/media/service/MediaServiceTest.kt
@@ -8,7 +8,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.engine.media.Media
-import mozilla.components.feature.media.notification.MediaNotificationFeature
+import mozilla.components.feature.media.MediaFeature
 import mozilla.components.feature.media.state.MediaStateMachine
 import mozilla.components.feature.media.state.MockMedia
 import mozilla.components.support.test.any
@@ -43,7 +43,7 @@ class MediaServiceTest {
 
         MediaStateMachine.start(sessionManager)
 
-        val feature = MediaNotificationFeature(mock())
+        val feature = MediaFeature(mock())
         feature.enable()
 
         media.playbackState = Media.PlaybackState.PLAYING
@@ -67,7 +67,7 @@ class MediaServiceTest {
 
         MediaStateMachine.start(sessionManager)
 
-        val feature = MediaNotificationFeature(mock())
+        val feature = MediaFeature(mock())
         feature.enable()
 
         media.playbackState = Media.PlaybackState.PLAYING
@@ -97,7 +97,7 @@ class MediaServiceTest {
 
         MediaStateMachine.start(sessionManager)
 
-        val feature = MediaNotificationFeature(mock())
+        val feature = MediaFeature(mock())
         feature.enable()
 
         media.playbackState = Media.PlaybackState.PLAYING
